requirements.txt
Flask
python-dotenv
pypdf
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
huggingface-hub>=0.36.0
transformers